// src/models/detailMovieCasts.ts
export interface detailMovieCasts {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface detailMovieCrew {
  adult: boolean;
  credit_id: string;
  department: string;
  gender: number;
  id: number;
  job: string;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
}

export default interface detailMovieCastsInterface {
  id: number;
  cast: detailMovieCasts[];
  crew: detailMovieCrew[];
}